requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
pymannkendall
google-generativeai
statsmodels
scikit-learn
openpyxl